vm: mounts: rename to inotify
